First (Feature) Commit! Yay!
Description of Product:
A diabetic needs to draw blood in order to get a blood glucose reading.
This reading is necessary to determine the proper insulin dosage at any
given time. In order to help preserve one's fingers, it is important to
test more or less evenly, but it is quite difficult to test in a truly
random fashion, which is where this utility comes in. "Basically, Finger
Wheel is a tool that allows a user to keep track of their last tested
finger. There is also a set of options allowing the user to disable
certain fingers in case they can't use them for any reason. The name
"Finger Wheel" comes from the original, physical prototype, which was in
the shape of a wheel. It would spin to advance.

Features:
- Hand graphics with highlight, moved by pressing arrows
- Text indicator
- Settings for disabling fingers in case they become unusable
- Saving functionality for current indicator position and also settings
- "About" screen with description
- Intro screen with name and author name